langchain-learning/streamlit/requirements.txt
# ── Core ──────────────────────────────────────────────────────
streamlit
python-dotenv

# ── LangChain base ────────────────────────────────────────────
langchain
langchain-core
langchain-community

# ── LLM Providers ────────────────────────────────────────────
# All six providers are listed — install all to avoid import
# errors when switching keys. Each is only imported at runtime
# inside build_llm() when that provider is actually selected.
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-cohere
langchain-mistralai
langchain-groq

# ── Text To Speech (required for 6_TextToSpeech.py only) ─────
gtts